Update time_check.py
After assembling the full program, I noticed that the timestamp identifier for comparison gave many false positives, and the program was not fine tuned to the specific amount of swaps queried and gathered. A new function was created to compare the id's and query the exact amount of swaps that were unique identifiers, thus saving extreme overlap and successfully gathering the up to date swaps.